Change configuration options.
